requirements.txt
flask~=3.0.3
psycopg2-binary
Flask-SQLAlchemy
python-dotenv~=1.0.1